label changes, renaming, clean up, ...
