+++ src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/[configId]/tooltips.tsx
